requirements.txt
boto3==1.34.162
botocore==1.34.162
polars==1.5.0
pyarrow==16.1.0
pandas==2.2.2
zstandard==0.22.0
python-dateutil==2.9.0.post0
fsspec==2025.7.0
PyYAML==6.0.2
